% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIFS.R
\name{ggsave_ifs}
\alias{ggsave_ifs}
\title{Save a ggplot with IFS base dimensions}
\usage{
ggsave_ifs(
  filename,
  plot = last_plot(),
  path = getwd(),
  widescreen = FALSE,
  save_data = FALSE,
  data_dir = path,
  data_name = filename,
  height = NA,
  width = NA,
  units = "mm",
  dpi = 320,
  device = "png",
  ...
)
}
\arguments{
\item{filename}{Filename to create on disk (including .png)}

\item{plot}{Plot to save}

\item{path}{Path of the directory to save plot to: path and filename are
combined to create the fully qualified file name. Defaults to current working directory.}

\item{widescreen}{Default plot size is 100mm x 150mm. If \code{TRUE}, saves to 100mm x 60mm.}

\item{save_data}{If \code{TRUE}, then it saves the data associated with the ggplot as a .csv file. This is not recommended for histograms / density plots.}

\item{data_dir}{The directory to store the data if \code{save_data == TRUE}. The default is the directory entered in \code{path}.}

\item{data_name}{The directory to}

\item{height}{The height of the plot (default in mm)}

\item{width}{The width of the plot (default in mm)}

\item{units}{The units for height and width (can take one of "mm", "in", "cm", or "px")).}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.}

\item{device}{Device to use, defaults to ".png". Can also be one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{...}{Any other arguments for \code{ggsave}, such as changing heights and widths, dpi, format.}
}
\value{

}
\description{
This is a wrapper for \code{ggsave} which saves ggplots in default IFS
dimensions for Word documents.
}
\examples{
\dontrun{g <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
ggsave_ifs("a_graph.png", g, "~/output")}
}
