% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIFS.R
\name{ifs_colour_wheel}
\alias{ifs_colour_wheel}
\title{Title Return hex codes for a vector of IFS colour names}
\usage{
ifs_colour_wheel(values)
}
\arguments{
\item{values}{A vector of colour names (e.g. "mid_green", "dark_green_2", "light_purple_1")}
}
\value{

}
\description{
Title Return hex codes for a vector of IFS colour names
}
\examples{
ifs_colour_wheel(c("dark_green_1", "mid_green", "dark_red_1"))
}
