% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIFS.R
\name{ifs_pal_finder}
\alias{ifs_pal_finder}
\title{Title Internal function for concatenating the colour list}
\usage{
ifs_pal_finder(direction, desired_cols)
}
\arguments{
\item{direction}{Whether colours should be reversed (-1) or not.}

\item{desired_cols}{The list of desired colours}
}
\value{

}
\description{
Title Internal function for concatenating the colour list
}
