% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIFS.R
\name{ifs_palettes}
\alias{ifs_palettes}
\title{Title Find the hex codes associated with a palette name}
\usage{
ifs_palettes(palette = "default")
}
\arguments{
\item{palette}{The palette (can take one of "default", "greens", "greys", "reds", "blues", "yellows" or "purples") the user wants.}
}
\value{

}
\description{
Title Find the hex codes associated with a palette name
}
\examples{
ifs_palettes("greens")
}
