% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIFS.R
\name{scale_colour_ifs}
\alias{scale_colour_ifs}
\alias{scale_fill_ifs}
\title{Title Add IFS colours to a ggplot
A wrapper for \code{scale_colour_manual} / \code{scale_fill_manual} that allows the user to specify a
certain colour palette, or ask for specific IFS colours.}
\usage{
scale_colour_ifs(palette = "default", values = NA, gradient = FALSE, ...)

scale_fill_ifs(palette = "default", values = NA, gradient = FALSE, ...)
}
\arguments{
\item{palette}{The default option are the IFS mid-colours. Accepted values are \code{"greens",
"greys", "blues", "reds", "yellows", "purples"}.}

\item{values}{For manually entering colour values. The allowed values are \code{"dark_X_2", "dark_X_1", "mid_X", "light_X_1", "light_X_2", "light_X_3"}.
Replace X with green, grey, blue, red, yellow, purple.}

\item{gradient}{If true, acts as a wrapper for \code{scale_colour_gradientn} or \code{scale_fill_gradientn}.}

\item{...}{Other values to pass to \code{scale_colour_manual} or \code{scale_fill_manual}}
}
\value{

}
\description{
Title Add IFS colours to a ggplot
A wrapper for \code{scale_colour_manual} / \code{scale_fill_manual} that allows the user to specify a
certain colour palette, or ask for specific IFS colours.
}
\section{Functions}{
\itemize{
\item \code{scale_fill_ifs}: Equivalent wrapper for scale_fill_manual
}}

\examples{
ggplot2::ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, colour = Species)) +
geom_point() +
scale_colour_ifs()

ggplot2::ggplot(iris, aes(x = Sepal.Width, y= Sepal.Length, colour = Species)) +
geom_point() +
scale_colour_ifs(values = c("dark_green_2", "mid_red", "light_yellow_3"))

ggplot2::ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, colour = Petal.Length)) +
geom_point() +
scale_colour_ifs(palette = "greys", gradient = TRUE)
}
