% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIFS.R
\name{theme_ifs}
\alias{theme_ifs}
\title{Add the IFS graph theme}
\usage{
theme_ifs(base_size = 10, ...)
}
\arguments{
\item{base_size}{Base font size, given in pts.}
}
\value{

}
\description{
\code{theme_ifs} adds the standard IFS graph theme, (but not the
IFS colour scheme) to a ggplot graph. \code{theme_ifs} can be overwritten
by personal theme preferences below the \code{theme_ifs} call.
}
\examples{
ggplot(mtcars, aes(x = disp, y = mpg)) +
theme_ifs()
}
